Adding row update logic with some animation.
Used to reload the entire table in a block supplied to the station
choice view controller. That's technically incorrect since the table
might not be loaded (is nil) due to low memory. The correct way is to
update the table when the control comes back to the main view
controller. The update is now done in the viewDidAppear.

TODO:
Not sure if animation is necessary. Design decision on that needs to be
made.