Adding map annotation for the selected source station.
When the user selects the source bart station, the annotation which
points to the station' location will be added to the map.

Note that there is only one annotation at a time.
When the user changes the source station, the existing annotation will be updated to point to the
newly selected station.