Adding logic to check if the updated user location should be processed.
Currently the map view updates the user location periodically, and it
seems like we don't really have controll over it.

So I added logic to check if the latest updated location is farther
than the previously recorded user location by 200m.

If not, we simply don't do anything and
don't record the updated location. If it is farther than 200m, the app
processes the location update and refetches the nearby stations.